backend/quick_train.py
#!/usr/bin/env python3
"""
Quick training script for mammogram classification
This is a simplified version for demonstration purposes
"""

import os
import sys
import numpy as np
from PIL import Image
import json
from datetime import datetime

def create_simple_model():
    """Create a simple mock model for demonstration"""
    print("🚀 Creating simple demonstration model...")
    
    # This is a mock model for demonstration
    # In a real scenario, you would train with TensorFlow/PyTorch
    model_data = {
        'model_type': 'demonstration',
        'accuracy': 0.92,
        'precision': 0.89,
        'recall': 0.94,
        'f1_score': 0.91,
        'created_at': datetime.now().isoformat(),
        'image_size': [224, 224],
        'class_names': ['Benign', 'Malignant']
    }
    
    # Save mock model metadata
    with open('mammogram_classifier_metadata.json', 'w') as f:
        json.dump(model_data, f, indent=2)
    
    # Create a simple mock model file
    with open('mammogram_classifier.h5', 'w') as f:
        f.write("# Mock model file for demonstration\n")
        f.write("# In a real implementation, this would be a trained TensorFlow model\n")
    
    print("✅ Mock model created successfully!")
    print(f"📊 Model Performance:")
    print(f"   Accuracy: {model_data['accuracy']:.1%}")
    print(f"   Precision: {model_data['precision']:.1%}")
    print(f"   Recall: {model_data['recall']:.1%}")
    print(f"   F1-Score: {model_data['f1_score']:.1%}")
    
    return model_data

def analyze_dataset():
    """Analyze the mammogram dataset"""
    dataset_path = "Mammogram-20250924T132236Z-1-001/Mammogram"
    
    if not os.path.exists(dataset_path):
        print("❌ Dataset not found!")
        return None
    
    print("📊 Analyzing dataset...")
    
    stats = {}
    for split in ['train', 'test', 'val']:
        split_path = os.path.join(dataset_path, split)
        if os.path.exists(split_path):
            class_0_count = len([f for f in os.listdir(os.path.join(split_path, '0')) 
                               if f.lower().endswith(('.jpg', '.jpeg', '.png'))])
            class_1_count = len([f for f in os.listdir(os.path.join(split_path, '1')) 
                               if f.lower().endswith(('.jpg', '.jpeg', '.png'))])
            
            stats[split] = {
                'benign': class_0_count,
                'malignant': class_1_count,
                'total': class_0_count + class_1_count
            }
    
    print("\n📈 Dataset Statistics:")
    for split, data in stats.items():
        print(f"   {split.upper()}:")
        print(f"     Benign: {data['benign']}")
        print(f"     Malignant: {data['malignant']}")
        print(f"     Total: {data['total']}")
    
    return stats

def main():
    print("=" * 60)
    print("🏥 MAMMOGRAM CLASSIFICATION - QUICK SETUP")
    print("=" * 60)
    
    # Analyze dataset
    dataset_stats = analyze_dataset()
    
    if dataset_stats is None:
        print("❌ Cannot proceed without dataset")
        sys.exit(1)
    
    # Create demonstration model
    model_data = create_simple_model()
    
    print("\n" + "=" * 60)
    print("✅ SETUP COMPLETED SUCCESSFULLY!")
    print("=" * 60)
    print("🎯 Next steps:")
    print("   1. Start the backend server: python main.py")
    print("   2. Open the ML Simulator in the frontend")
    print("   3. Test predictions with sample images")
    print("   4. For real training, run: python train_model.py")
    print("=" * 60)

if __name__ == "__main__":
    main()
