backend/requirements.txt
fastapi
uvicorn
pydantic
email-validator
pandas
numpy
scikit-learn
tensorflow
pillow
python-multipart==0.0.6
tensorflow>=2.16.1
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0
numpy>=1.24.3
openpyxl>=3.1.0
joblib>=1.3.0
pandas>=2.0.0
